chore: moved eventSource for sharing
