+++ packages/sdk/cloudflare/src/platform/requests.ts
@@ -1,3 +1,4 @@
+import { MockEventSource } from '@launchdarkly/js-server-sdk-common';
 import type {
   EventSource,
   EventSourceInitDict,
@@ -5,7 +6,6 @@ import type {
   Response,
   Requests,
 } from '@launchdarkly/js-server-sdk-common';
-import MockEventSource from './eventSource';
 
 export default class CloudflareRequests implements Requests {
   fetch(url: string, options: Options = {}): Promise<Response> {
