+++ packages/shared/sdk-server/src/events/index.ts
@@ -1,0 +1,4 @@
+import MockEventSource from './MockEventSource';
+
+// eslint-disable-next-line import/prefer-default-export
+export { MockEventSource };
