+++ packages/shared/sdk-server/src/index.ts
@@ -5,6 +5,7 @@ export * as integrations from './integrations';
 export * as platform from '@launchdarkly/js-sdk-common';
 export * from './api';
 export * from './store';
+export * from './events';
 export * from '@launchdarkly/js-sdk-common';
 
 export { LDClientImpl, BigSegmentStoreStatusProviderImpl };
